fix: 스토리북 데코레이터를 워크샵에서 관리하도록 수정
