+++ apps/web/src/app/(onboarding)/onboarding/_components/OnboardingHeader/OnboardingHeader.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { Step, Text } from '@vook-client/design-system'
 
 import { onboardingHeader, stepArea } from './OnboardingHeader.css'
