+++ apps/web/src/components/SignUpForm/SignUpForm.stories.tsx
@@ -1,9 +1,4 @@
-import {
-  AppRouterContext,
-  type AppRouterInstance,
-} from 'next/dist/shared/lib/app-router-context.shared-runtime'
 import type { Meta, StoryObj } from '@storybook/react'
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
 import { SignUpForm } from './SignUpForm'
 
@@ -13,16 +8,11 @@ const meta: Meta<typeof SignUpForm> = {
     layout: 'centered',
   },
   component: SignUpForm,
-  tags: ['autodocs'],
   decorators: [
     (Story) => (
-      <AppRouterContext.Provider value={{} as AppRouterInstance}>
-        <QueryClientProvider client={new QueryClient()}>
-          <div style={{ width: 400 }}>
-            <Story />
-          </div>
-        </QueryClientProvider>
-      </AppRouterContext.Provider>
+      <div style={{ width: 400 }}>
+        <Story />
+      </div>
     ),
   ],
 }
