+++ apps/workshop/.storybook/manager.ts
@@ -1,5 +1,4 @@
 import { addons } from '@storybook/manager-api'
-import { themes } from '@storybook/theming'
 import { create } from '@storybook/theming/create'
 
 addons.setConfig({
