+++ apps/workshop/package.json
@@ -9,8 +9,10 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@tanstack/react-query": "^5.32.0",
     "@vook-client/api": "*",
     "@vook-client/design-system": "*",
+    "next": "^14.1.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
